problems/src/test/java/com/abotnaru/problems/neetcode/graphs/MaxAreaIslandTest.java
package com.abotnaru.problems.neetcode.graphs;

import static org.junit.jupiter.api.Assertions.assertEquals;

import org.junit.jupiter.api.Test;

class MaxAreaIslandTest {

	@Test
	void testMaxAreaOfIslandExpectSix() {
		int[][] grid = new int[][] { { 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0 },
				{ 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0 }, { 0, 1, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0 },
				{ 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0 }, { 0, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 0, 0 },
				{ 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0 }, { 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0 },
				{ 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0 } };
		assertEquals(6, new MaxAreaIsland().maxAreaOfIsland(grid));
	}

	@Test
	void testMaxAreaOfIslandExpectZero() {
		int[][] grid = new int[][] { { 0, 0, 0, 0, 0, 0, 0, 0 } };
		assertEquals(0, new MaxAreaIsland().maxAreaOfIsland(grid));
	}

}
